Refactor database migration tooling
