+++ .env.example
@@ -13,7 +13,9 @@ SESSION_SECRET=${SECRET_KEY}
 # --- Base de datos (PostgreSQL con psycopg 3) ---
 # Ejemplo DEV local con contenedor en 5433:
 # docker run --name obyra-pg -p 5433:5432 -e POSTGRES_USER=obyra -e POSTGRES_PASSWORD=obyra -e POSTGRES_DB=obyra_dev -d postgres:16
-DATABASE_URL=postgresql+psycopg://obyra:obyra@localhost:5433/obyra_dev
+DATABASE_URL=postgresql+psycopg://USER:PASS@localhost:5433/obyra_dev
+ALEMBIC_DATABASE_URL=postgresql+psycopg://MIGRATOR_USER:MIGRATOR_PASS@localhost:5433/obyra_dev
+READONLY_DATABASE_URL=postgresql+psycopg://READONLY_USER:READONLY_PASS@localhost:5433/obyra_dev
 SQLALCHEMY_POOL_SIZE=10
 SQLALCHEMY_MAX_OVERFLOW=10
 SQLALCHEMY_POOL_TIMEOUT=30
