+++ alembic/versions/0001_initial_baseline.py
@@ -1,0 +1,19 @@
+"""Revisión base sin operaciones destructivas.
+
+Esta revisión sirve como punto de partida para futuras
+migraciones estructurales dentro del esquema ``app``.
+"""
+
+# revision identifiers, used by Alembic.
+revision = "0001_initial_baseline"
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade() -> None:
+    pass
+
+
+def downgrade() -> None:
+    pass
