+++ extensions.py
@@ -1,6 +1,5 @@
 from flask_sqlalchemy import SQLAlchemy
 from flask_login import LoginManager
-from flask_migrate import Migrate
 from sqlalchemy.orm import DeclarativeBase
 
 
@@ -10,4 +9,3 @@ class Base(DeclarativeBase):
 
 db = SQLAlchemy(model_class=Base)
 login_manager = LoginManager()
-migrate = Migrate()
