+++ requirements.txt
@@ -3,8 +3,8 @@
 authlib~=1.6.5
 flask~=3.1.1
 flask-login~=0.6.3
-flask-migrate~=4.0.7
 flask-sqlalchemy~=3.1.1
+alembic~=1.13.3
 gunicorn~=23.0.0
 jinja2~=3.1.4
 matplotlib~=3.10.5
